feat(supervised_install): Switch to deb installer
BREAKING CHANGE: The Home Assistant Supervised install.sh convenience script is deprecated.
As a replacement, the install is now performed by installing the homeassistant-supervised.deb package.
Overriding the default values for many of the  configurations is not currently possible with the .deb package install, so they have been removed from the role.

Fixes #27